Add more tests

+++ test.ts
@@ -1,12 +1,14 @@
 import kleur from "npm:kleur";
 
-async function test(file: string) {
+async function test(file: string, keep?: boolean) {
   console.log("");
   console.log(kleur.green().inverse(file));
-  try {
-    Deno.removeSync("jieba-dyn.so");
-  } catch (_e) {
-    // Ignore
+  if (!keep) {
+    try {
+      Deno.removeSync("jieba-dyn.so");
+    } catch (_e) {
+      // Ignore
+    }
   }
   const process = Deno.run({
     cmd: ["cask", "emacs", "--batch", "-l", file],
@@ -14,15 +16,25 @@ async function test(file: string) {
   const status = await process.status();
   if (status.success) {
     console.log(kleur.green(`${file} ran without error`));
+  } else {
+    console.log(kleur.red(`${file} failed`));
   }
 }
 
-for (const file of [
-  "tests/require-download.el",
-  "tests/require-compile.el",
-  "tests/require-directly.el",
-]) {
-  await test(file);
+// Specify file -> run those tests
+// Don't -> we're testing requring the library
+if (Deno.args.length > 0) {
+  for (const file of Deno.args) {
+    await test(file, true);
+  }
+} else {
+  for (const file of [
+    "tests/require-download.el",
+    "tests/require-compile.el",
+    "tests/require-directly.el",
+  ]) {
+    await test(file);
+  }
 }
 
 // Local Variables:
